requirements.txt
certifi==2025.6.15
charset-normalizer==3.4.2
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
decorator==5.2.1
ffmpeg-python==0.2.0
filelock==3.18.0
filetype==1.2.0
fonttools==4.58.4
fsspec==2025.5.1
future==1.0.0
idna==3.7
imageio==2.37.0
imageio-ffmpeg==0.6.0
Jinja2==3.1.6
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.3
moviepy==2.2.1
mpmath==1.3.0
networkx==3.5
numpy==2.3.1
opencv-python==4.11.0.86
opencv-python-headless==4.10.0.84
packaging==25.0
pandas==2.3.0
pillow==11.2.1
pillow-avif-plugin==1.5.2
pillow_heif==1.0.0
proglog==0.1.12
psutil==7.0.0
py-cpuinfo==9.0.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
requests-toolbelt==1.0.0
roboflow==1.2.0
scipy==1.16.0
setuptools==80.9.0
six==1.17.0
sympy==1.14.0
torch==2.7.1+cu128
torchaudio==2.7.1+cu128
torchvision==0.22.1+cu128
tqdm==4.67.1
typing_extensions==4.14.0
tzdata==2025.2
ultralytics==8.3.160
ultralytics-thop==2.0.14
urllib3==2.5.0
